# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MagicLinkRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "turnstile_token": {
            "title": "Turnstile Token",
            "type": "string"
          }
        },
        "required": [
          "email",
          "turnstile_token"
        ],
        "title": "MagicLinkRequest",
        "type": "object"
      },
      "MagicLinkSent": {
        "properties": {
          "status": {
            "default": "sent",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "MagicLinkSent",
        "type": "object"
      },
      "MagicLinkVerify": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "MagicLinkVerify",
        "type": "object"
      },
...
      "SignInResult": {
        "properties": {
          "next": {
            "title": "Next",
            "type": "string"
          }
        },
        "required": [
          "next"
        ],
        "title": "SignInResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cornerpin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/google/callback": {
      "get": {
        "operationId": "google_callback_v1_auth_google_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "cookie",
            "name": "cp_oauth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Oauth"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/google/start": {
      "get": {
        "operationId": "google_start_v1_auth_google_start_get",
        "parameters": [
          {
            "in": "query",
            "name": "next",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Next"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Start",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/magic-link": {
      "post": {
        "description": "Always 202 for a human, whether or not the address has an account.",
        "operationId": "request_magic_link_v1_auth_magic_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkSent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Magic Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/magic-link/verify": {
      "post": {
        "operationId": "verify_magic_link_v1_auth_magic_link_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Expired or used link"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Magic Link",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/signout": {
      "post": {
        "operationId": "sign_out_v1_auth_signout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign Out",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}